Partial commit. Mostly layout changes.
The content viewer's layout has changed so that the right chat panel does not
sit under the content bar, but to the right of everything.

Some styling changes have also been made.

(And also caught up in this code to make the polls disable checkboxes when
you have selected as many options as allowed)