Added activity stream / newsfeed
This is a social media style newsfeed, that lists added content entries.

Longer term, ContentEntry will gain a few fields: published,
newsworthy, and list.  (To control whether something appears in the
newsfeed, in the content index, and whether or not it's "published"
[because many items are created before they are edited])

Technically, this commit also changes how the main content views are
inserted.  There is now an ng-include that takes its template URL from
a controller, rather than a dirty great ng-switch.  This means that the
template for a particular kind of content item is loaded as needed,
rather than them all being loaded at the start.