ContentController returns the JSON of the entry
Previously there had been an error where editItem was returning it as { entry: json } and editTags was returning it as { course: json }.

This commit also makes the MarkdownPage view controller watch for changes to the item text, rather than just rendering once when it is first instantiated.

Between them, this fixes some buggy editing behaviour for content entries.