Removed ng-app attribute that was causing Angular.js to bootstrap before scripts were loaded
Angular.js bootstraps automatically if it sees an ng-app attribute on the page.

However, as our scripts load through require.js, we bootstrap Angular.js from there (in min.js) instead

The line to bootstrap Angular.js from min.js was already in place, but there had been an ng-app tag in the template, causing Angular to try to bootstrap automatically as well.