Move the Impressory scripts after the directive templates in templateFront
One of our users found that sometimes the directives were not rendering, and the console was full of failed calls to the server to fetch the templates.  This was an intermittent error.  It occurred to me that it may be a race condition between the scripts and the browser's html parsing.  As Angular.js was being initialised earlier in the page than the script elements that contain the templates for the directives, perhaps on occasion Angular.js would look for the template before the browser's html parsing had reached the script element containing it.

This commit moves the require.js invocation of Impressory to the last element in the page, so that the script elements containing the directive templates will already have been parsed by the browser before Angular.js is invoked.