FindById.findMany returns items in the same order as the sequence of keys
The request to the database asks for items where the id is in a particular set.  If there are duplicate entries, this only returns the item once, and the returned items are not necessarily in the same order as the request.

Previously the returned items were then simply returned to the caller -- meaning the result could be out of order and with duplicates removed.

findMany now collects the results from the database, and then returns them in-order, including any duplicates, as requested.