Cache the request's LazyId, but don't automatically look it up
There was a curious bug happening on the AWS server (but not reproduce able elsewhere, even when running the same JDK etc in vagrant), where if the user was logged in, the entire Play app would freeze.  Still not known what was causing this problem, but this change both fixes it (for now) and is also sensible anyway -- there may be requests that create an Approval but do not need to load the User from the database (for instance if the only permissions it asks for are always approved).