Upgrade to handy-appbase, play2.2.  Partial commit.
This is the first commit in a fairly large set of changes.

The major change here is using handy-appbase (which gives us DataAction and allows us to remove a lot of duplicated code)

Some JSON APIs have changed -- particularly returning the object or array directly, rather than wrapped in an outer object. For example, returning an array of ContentEntries, rather than { entries -> [ ... ] }

The beginnings of a plug-in architecture are also included, as new ContentItems can register the handlers we need for their JSON and URLs.