Polls now fetch the vote you've already made from the database. Added EventViews
This commit has a small amount of tweaking done in the process of neatening how polls work.

EventViews is introduced, as a registry of Angular.js partial templates for rendering events in the chat stream.

The poll javascript has its package renamed to include 'plugins', as a prelude to eventually moving it out into impressory-poll

PushPollToChat now just contains the poll id, not the JSON for the poll itself. The client will load the poll (and it may already be in the ContentService's cache)

ContentItemToJson now passes the entry as well as the item into toJsonFor, in case the process of converting it to JSON needs to look anything up using the entry (e.g., the entry ID or course)