Add content uses an ng-if for share section, so that item stream template is only loaded after the item is set
Previously there was a problem where the preview of a web page when adding it would always use the "default" stream template, rather than the web page stream template. This was because the directive to render the share item was present before there was an item to share, so it would try to load a template for "undefined" and would not change it when the item became present.