handy-0.5-RC1 makes a distinction whether a Ref's id is immediately available or not.
Previously, calls to Ref.getId could be blocking. For instance, request.user is a RefFuture looking up the user by their session key, so a call to getId required the Future to complete.

Changes to handy introduce the Ref.refId, which accesses a Ref's ID asynchronously.  RefWithId (Ref with IdImmediate) were also introduced, marking those Ref where the ID is immediately available and there is a getId method.

This commit updates Impressory accordingly.